Proves Enquesta + insert or update

+++ src/main/java/com/tecnocampus/useCases/PreguntaCasosUs.java
@@ -2,6 +2,9 @@ package com.tecnocampus.useCases;
 
 import com.tecnocampus.BeansManager;
 import com.tecnocampus.domain.*;
+import com.tecnocampus.exceptions.PreguntaNoExisteixException;
+import com.tecnocampus.exceptions.RespostaDuplicadaException;
+import com.tecnocampus.exceptions.RespostaForaDeLimitsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,17 +32,30 @@ public class PreguntaCasosUs {
 
     public Resposta afegirResposta(Pregunta pregunta, Usuari usuari, int valor) throws RuntimeException {
 
-        if(usuari == null) throw new IllegalArgumentException("usuari");
-        if(pregunta == null) throw new IllegalArgumentException("pregunta");
+        if(usuari == null)
+            throw new IllegalArgumentException("usuari");
+        if(pregunta == null)
+            throw new IllegalArgumentException("pregunta");
 
         // Un usuari no pot respondre la mateixa pregunta 2 cops
         if (!beansManager.respostaRepository.canAnswer(pregunta, usuari))
-            throw new RuntimeException("Un usuari no pot respondre la mateixa pregunta 2 cops");
+            throw new RespostaDuplicadaException("Un usuari no pot respondre la mateixa pregunta 2 cops");
+
+        if(valor < ((PreguntaNumerica)pregunta).getMinim())
+            throw new RespostaForaDeLimitsException("El valor de la resposta no pot ser inferior el valor mínim de la pregunta");
+        if(valor > ((PreguntaNumerica)pregunta).getMaxim())
+            throw new RespostaForaDeLimitsException("El valor de la resposta no pot superar el valor màxim de la pregunta");
 
         RespostaNumerica resposta = new RespostaNumerica();
-        resposta.setUsuari(usuari);
+
         resposta.setValor(valor);
+
+        resposta.setUsuari(usuari);
+        usuari.afegirResposta(resposta);
+
         pregunta.afegirResposta(resposta);
+        resposta.setPregunta(pregunta);
+
         beansManager.respostaRepository.save(pregunta, resposta);
         return resposta;
 
@@ -51,8 +67,10 @@ public class PreguntaCasosUs {
      * @return
      * @throws Exception
      */
-    public int eliminarPregunta(Pregunta pregunta) throws Exception {
-        return beansManager.preguntaRepository.delete(pregunta);
+    public void eliminarPregunta(Pregunta pregunta) throws RuntimeException {
+        if (beansManager.preguntaRepository.findOne(pregunta.getId()) != null)
+            throw new PreguntaNoExisteixException();
+        beansManager.preguntaRepository.delete(pregunta);
     }
 
     /***
@@ -60,7 +78,7 @@ public class PreguntaCasosUs {
      * @param enquesta
      * @throws Exception
      */
-    public void eliminarTotesPreguntes(Enquesta enquesta) throws Exception {
+    public void eliminarTotesPreguntes(Enquesta enquesta) throws RuntimeException {
         for (Pregunta pregunta: enquesta.getPreguntes()) {
             eliminarPregunta(pregunta);
         }
